beyondcompare: use Get-UICulture to detect language
Get-Culture returns settings such as the keyboard layout, and the
display format of items such as numbers, currency, and dates.
Get-UICulture returns the language chosen for the user interface. It is
very common to use different values for these two settings. As a result,
it can happen that a translated version of the application is installed
even though the rest of Windows is in English.